Store server.jar and data in separate folders
Reason for this is to be able to store data and server config file in a volume, so that this data can be made persistent and not live inside the container.

The problem with the current setup is if /data is setup as a volume mapped to a host folder, that folder overwrites the container's /data, and thereby also server.jar which was stored in that folder. 
The server.properties config file needs to be in the same folder as server.jar, but a symbolic link makes that happen and the server.properties can now be stored with the data.

If one cheeses not to map /data to a volume, this docker will work the same as it used to.